requirements.txt
flask
flask-cors
gunicorn
stockfish
requests
python-chess
google-generativeai
dotenv